requirements.txt
peewee==3.17.0
paho-mqtt==1.6.1
flask==2.3.3
flask-cors==4.0.0